Retire rails-assets
All rails-assets dependencies were moved to a package.json file, and
now they are managed by NPM.

All javascript files are now bundled together with the help of Webpack
and its babel loader. CSS from Javascript libraries are now also
included in the Webpack bundle and loaded into a style tag with the help
of the style-loader.