app/assets/javascripts/components refactor
This commit is about folders reorganization and files renaming - instead
of naming 'foo_chart.es6.jsx' all componentes that represent a chart,
put them in a /charts folder, for example.

I also defined the difference between /panels and /controls folders: the
first holds all componentes that are containers, and the last contains
all components that in facts change the state of the application.